refactor: extract GlobalUndoToast from App.tsx
