+++ spa/src/App.tsx
@@ -1,5 +1,5 @@
 // spa/src/App.tsx — v2 重構：wouter Router + Tab/Pane model
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { useCallback, useEffect, useMemo, useState } from 'react'
 import { Router } from 'wouter'
 import { ActivityBar } from './components/ActivityBar'
 import { TabBar } from './components/TabBar'
@@ -18,7 +18,6 @@ import { useShortcuts } from './hooks/useShortcuts'
 import { openBrowserTab } from './lib/open-browser-tab'
 import './lib/browser-shortcuts'
 import { useNotificationDispatcher } from './hooks/useNotificationDispatcher'
-import { useUndoToast } from './stores/useUndoToast'
 import { useTabWorkspaceActions } from './hooks/useTabWorkspaceActions'
 import { isStandaloneTab } from './types/tab'
 import {
@@ -32,42 +31,8 @@ import { RenamePopover } from './components/RenamePopover'
 import { ThemeInjector } from './components/ThemeInjector'
 import { ErrorBoundary } from './components/ErrorBoundary'
 import { getPlatformCapabilities } from './lib/platform'
-import { useI18nStore } from './stores/useI18nStore'
 import type { Tab } from './types/tab'
-
-function GlobalUndoToast() {
-  const toast = useUndoToast((s) => s.toast)
-  const dismiss = useUndoToast((s) => s.dismiss)
-  const t = useI18nStore((s) => s.t)
-  const timerRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-
-  useEffect(() => {
-    if (!toast) return
-    timerRef.current = setTimeout(() => dismiss(), 5000)
-    return () => {
-      if (timerRef.current) clearTimeout(timerRef.current)
-    }
-  }, [toast, dismiss])
-
-  if (!toast) return null
-
-  return (
-    <div className="fixed bottom-4 left-1/2 -translate-x-1/2 bg-zinc-800 border border-zinc-700 rounded-lg px-4 py-3 flex items-center gap-3 shadow-lg z-50">
-      <span className="text-sm text-zinc-300">
-        {toast.message}
-      </span>
-      <button
-        className="text-sm text-blue-400 hover:text-blue-300 font-medium cursor-pointer"
-        onClick={() => {
-          toast.restore()
-          dismiss()
-        }}
-      >
-        {t('hosts.undo')}
-      </button>
-    </div>
-  )
-}
+import { GlobalUndoToast } from './components/GlobalUndoToast'
 
 export default function App() {
   const isElectron = getPlatformCapabilities().isElectron
