+++ spa/src/components/GlobalUndoToast.tsx
@@ -1,0 +1,37 @@
+import { useEffect, useRef } from 'react'
+import { useUndoToast } from '../stores/useUndoToast'
+import { useI18nStore } from '../stores/useI18nStore'
+
+export function GlobalUndoToast() {
+  const toast = useUndoToast((s) => s.toast)
+  const dismiss = useUndoToast((s) => s.dismiss)
+  const t = useI18nStore((s) => s.t)
+  const timerRef = useRef<ReturnType<typeof setTimeout> | null>(null)
+
+  useEffect(() => {
+    if (!toast) return
+    timerRef.current = setTimeout(() => dismiss(), 5000)
+    return () => {
+      if (timerRef.current) clearTimeout(timerRef.current)
+    }
+  }, [toast, dismiss])
+
+  if (!toast) return null
+
+  return (
+    <div className="fixed bottom-4 left-1/2 -translate-x-1/2 bg-zinc-800 border border-zinc-700 rounded-lg px-4 py-3 flex items-center gap-3 shadow-lg z-50">
+      <span className="text-sm text-zinc-300">
+        {toast.message}
+      </span>
+      <button
+        className="text-sm text-blue-400 hover:text-blue-300 font-medium cursor-pointer"
+        onClick={() => {
+          toast.restore()
+          dismiss()
+        }}
+      >
+        {t('hosts.undo')}
+      </button>
+    </div>
+  )
+}
